third_party/blink/web_tests/external/wpt/app-history/navigate-event/navigate-form.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form id="form" action=""></form>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_equals(e.navigationType, "replace");
    assert_true(e.cancelable);
    assert_true(e.canRespond);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.destination.url, location.href + "?");
    assert_false(e.destination.sameDocument);
    assert_equals(e.destination.key, null);
    assert_equals(e.destination.id, null);
    assert_equals(e.destination.index, -1);
    assert_not_equals(e.formData, null);
    e.preventDefault();
  });
  window.onload = t.step_func(() => form.submit());
}, "<form> submission fires navigate event");
</script>
